Delivery app | move class to the right package
