+++ client_delivery/shared/src/commonMain/kotlin/data/gateway/remote/IdentityRemoteGateway.kt
@@ -1,8 +1,9 @@
 package data.gateway.remote
 
+import data.remote.mapper.toDto
 import data.remote.mapper.toEntity
 import data.remote.model.BaseResponse
-import data.remote.model.DeliveryRequestPermission
+import domain.entity.DeliveryRequestPermission
 import data.remote.model.SessionDto
 import domain.entity.Session
 import domain.gateway.remote.IIdentityRemoteGateway
@@ -49,11 +50,11 @@ class IdentityRemoteGateway(client: HttpClient) : IIdentityRemoteGateway,
     }
 
     override suspend fun createRequestPermission(
-        deliveryRequestPermission:DeliveryRequestPermission
+        deliveryRequestPermission: DeliveryRequestPermission
     ): Boolean {
         val result = tryToExecute<BaseResponse<Boolean>> {
             post("/restaurant-permission-request") {
-              setBody(deliveryRequestPermission)
+              setBody(deliveryRequestPermission.toDto())
             }
         }.value ?: throw Exception()
